feat: balance checker with shared logic used in both expiration detector and swap tracker which checks payment address balance and if it's different from what the daemon has detected as incoming amount, check incoming transactions over etherscan API
